Adding range calculator
Adding a class that, given a minimum and maximum number, finds all number ranges between the min and max (inclusive) such that a regular expression could be created that matches only the string representation of all numbers in each sub-range.  Only works with positive, base 10 integers for now. This is a first pass, could probably be more efficient.